# fix: Resolve JSONB+INET network filtering bug for production systems (#v0.5.7)