--- a/tests/core/test_production_fix_validation.py
+++ b/tests/core/test_production_fix_validation.py
@@ -29,10 +29,10 @@ class TestProductionFixValidation:
 
         # CRITICAL: The fix should now detect "8.8.8.8" as an IP and apply ::inet casting
         assert "::inet" in sql_str, f"Production fix failed - no inet casting: {sql_str}"
-        assert "host(" in sql_str, f"Should use host() function for IP comparison: {sql_str}"
+        # Note: Fixed behavior uses direct ::inet casting for equality, not host()
         assert "8.8.8.8" in sql_str, f"Should include IP address value: {sql_str}"
 
-        # The result should now be: host((data ->> 'ip_address')::inet) = '8.8.8.8'
+        # The result should now be: (data ->> 'ip_address')::inet = '8.8.8.8'
         # Instead of the broken: (data ->> 'ip_address') = '8.8.8.8'
 
     def test_various_ip_formats_detected_correctly(self):
@@ -179,10 +179,10 @@ class TestProductionFixValidation:
         # The fix converts this from broken text comparison:
         # (data ->> 'ip_address') = '8.8.8.8'
         # To working IP comparison:
-        # host((data ->> 'ip_address')::inet) = '8.8.8.8'
+        # (data ->> 'ip_address')::inet = '8.8.8.8'
 
-        # Verify we got the working version
-        assert "host(" in sql_str and "::inet" in sql_str, "Should use proper IP comparison"
+        # Verify we got the working version with proper INET casting
+        assert "::inet" in sql_str, "Should use proper IP comparison with INET casting"
 
 
 @pytest.mark.core