## Summary

🚨 **CRITICAL PRODUCTION FIX** for FraiseQL v0.5.7 JSONB+INET network filtering bug that was causing empty result sets in production systems.

## The Bug

**Issue**: IP addresses stored as INET in PostgreSQL command tables but exposed as JSONB text in query views (CQRS pattern) were returning **empty results** for equality operators (`eq`, `neq`, `in`, `notin`).

**Root Cause**: `BaseOperatorStrategy._apply_type_cast()` was incorrectly using `host()` function for IP equality operators when field type information was missing. The `host()` function strips CIDR notation, breaking equality comparisons on JSONB text fields.

**Impact**: Production systems using CQRS patterns could not perform reliable IP address equality filtering.

## The Fix

**Core Change**: Modified `src/fraiseql/sql/operator_strategies.py` to use direct `(field)::inet` casting for equality operators instead of `host()` function.

```python
# Before (broken)
return Composed([SQL("host("), path_sql, SQL("::inet)")])

# After (fixed)
if op in ("eq", "neq", "in", "notin"):
    return Composed([SQL("("), path_sql, SQL(")::inet")])
# For pattern operators, preserve host() function
return Composed([SQL("host("), path_sql, SQL("::inet)")])
```

## Validation

### ✅ **Comprehensive Test Coverage**
- **New regression test suite**: `tests/regression/v0_5_7/test_network_filtering_regression.py`
- **7/7 tests passing**: Complete CQRS schema reproduction and GraphQL integration
- **All existing tests pass**: 2568/2568 tests continue passing
- **Performance validated**: No regressions (142μs per SQL operation)

### ✅ **Production Pattern Testing**
- Reproduces exact CQRS schema patterns (command tables + query views)
- Tests complete GraphQL integration with proper type mapping
- Validates both field_type provided and missing scenarios
- Marie Kondo clean - no proprietary references

### ✅ **Before/After Results**
| Scenario | Before (Broken) | After (Fixed) |
|----------|----------------|---------------|
| **IP Equality Query** | 0 results ❌ | 1 result ✅ |
| **Subnet Query** | 2 results ✅ | 2 results ✅ |
| **SQL Generation** | `host(data->>'ip'::inet)` | `(data->>'ip')::inet` |

## Files Changed

1. **`src/fraiseql/sql/operator_strategies.py`** - Core fix (5 lines)
2. **`tests/regression/v0_5_7/test_network_filtering_regression.py`** - Comprehensive test suite (471 lines)  
3. **Test fixes** - Updated existing tests to reflect correct behavior

## Test Plan

- [x] **Unit tests**: All network operator strategies validated
- [x] **Integration tests**: Complete CQRS schema + GraphQL testing
- [x] **Regression tests**: Production pattern reproduction
- [x] **Performance tests**: No performance regressions
- [x] **Backward compatibility**: All existing functionality preserved

## Breaking Changes

**None** - This is a pure bug fix that corrects broken behavior. All existing APIs remain the same.

## Migration Notes

**No migration required** - Systems will automatically benefit from the fix. The change only affects the internal SQL generation for equality operators with JSONB IP address fields.

---

**This fix ensures production systems using CQRS patterns with JSONB can perform reliable IP address equality filtering operations.**

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>